+++ demo-tt/src/main/webapp/WEB-INF/classes/com/manydesigns/portofino/tt/Refresh.groovy
@@ -1,0 +1,5 @@
+package com.manydesigns.portofino.tt
+
+class Refresh {
+    boolean changeMe = true
+}